Modularize about page
