+++ src/components/about/AboutMission.jsx
@@ -1,0 +1,11 @@
+export default function AboutMission() {
+  return (
+    <section className="ab-numbers ab-mission" id="mission">
+      <div className="ab-mission__inner">
+        <span className="ab-mission__eyebrow">Our Mission</span>
+        <h2 className="ab-mission__statement">To show the world how beautiful <em>Tanzania</em> truly is.</h2>
+        <p className="ab-mission__sub">Simple, and the same as it was on that first piece of paper — every island, every coast, every corner of the mainland has a story worth meeting in person.</p>
+      </div>
+    </section>
+  );
+}
