+++ src/data/aboutPageData.js
@@ -1,0 +1,61 @@
+export const ABOUT_HERO_IMAGE = '/assets/images/safaris/crowned-cranes-in-grass.webp';
+export const ABOUT_STORY_IMAGE = '/assets/images/excursions/trips/dhow-heritage.webp';
+export const ABOUT_CTA_IMAGE = '/assets/images/excursions/dream-dhow-sunset.webp';
+
+export const aboutTimeline = [
+  {
+    year: 'Where it began',
+    title: 'A vision on the east coast of Zanzibar',
+    body: 'Years ago, while working in a hotel on the east coast of Zanzibar, the first idea appeared — to create something bigger than just tours or bookings. Something that would connect people to the beauty, culture and spirit of Tanzania. At the time it felt too big and too far away, but it never disappeared.',
+  },
+  {
+    year: 'The first steps',
+    title: 'Built together with family',
+    body: 'The company was registered together with family — the very first foundation of what would later become Destination Paradise. After a deep loss in those early days, everything slowed down. But the vision remained alive.',
+  },
+  {
+    year: 'The quiet years',
+    title: 'A dream that kept growing',
+    body: 'Through years of work, travel and other projects, the dream continued to grow little by little. What had once been written on a piece of paper kept asking to be built — patiently, in the background, waiting for its moment.',
+  },
+  {
+    year: 'A new chapter',
+    title: 'Building the platform from the ground up',
+    body: 'Moving to Germany opened a new chapter — learning coding, web development and media production, and gathering the technical skills needed to build the platform from the ground up. Slowly, an idea on paper became a real company with a real vision.',
+  },
+  {
+    year: 'Today',
+    title: 'Officially launching from Unguja',
+    body: 'After years of preparation, Destination Paradise is officially launching. We begin our journey in Unguja, Zanzibar and across mainland Tanzania — with Pemba and Mafia Island to follow.',
+  },
+];
+
+export const aboutPillars = [
+  {
+    icon: 'people',
+    title: 'Local drivers and guides',
+    body: 'We work hand-in-hand with drivers and guides from across the islands and mainland — the people who actually know the roads, the reefs and the stories.',
+  },
+  {
+    icon: 'home',
+    title: 'Local hotels and restaurants',
+    body: 'We partner with locally owned hotels and restaurants wherever we can. Travel money stays in the community that hosts you.',
+  },
+  {
+    icon: 'network',
+    title: 'A network, not a closed system',
+    body: 'Instead of building a closed operation, we are building a network — one that allows the wider community to grow together with us, step by step.',
+  },
+  {
+    icon: 'layers',
+    title: 'Authentic, designed by hand',
+    body: 'Every excursion, safari and cultural experience is designed carefully — so travellers meet Tanzania in a more authentic and meaningful way.',
+  },
+];
+
+export const aboutDestinations = [
+  { tag: 'Available now', name: 'Unguja, Zanzibar', body: 'Where the journey begins. Excursions, cultural experiences and trips across the main island.' },
+  { tag: 'Available now', name: 'Mainland Tanzania', body: 'The Serengeti, Ngorongoro, Kilimanjaro and beyond — the wider story of Tanzania, ready to be explored.' },
+  { tag: 'Next', name: 'Pemba Island', body: "Zanzibar's quieter sister — untouched reefs, clove forests, and a slower pace just across the channel." },
+  { tag: 'Next', name: 'Mafia Island', body: 'A marine park with whale sharks, turtles and some of the most pristine waters on the East African coast.' },
+];
