+++ src/styles/about/community.css
@@ -1,0 +1,152 @@
+/* ============ SUSTAINABILITY ============ */
+.ab-sus {
+  background: linear-gradient(180deg, rgba(184,84,26,.06) 0%, transparent 100%);
+  padding: clamp(4rem, 8vw, 7rem) clamp(1.25rem, 5vw, 6rem);
+}
+.ab-sus__inner { max-width: 1240px; margin: 0 auto; }
+.ab-sus__head {
+  display: grid; grid-template-columns: 1fr 1.2fr;
+  gap: clamp(2rem, 5vw, 4rem);
+  align-items: end;
+  margin-bottom: clamp(2.5rem, 5vw, 4rem);
+}
+@media (max-width: 880px) { .ab-sus__head { grid-template-columns: 1fr; align-items: start; } }
+.ab-sus__eyebrow { color: #b8541a !important; }
+.ab-sus__eyebrow::before { background: #b8541a !important; }
+.ab-sus__head h2 {
+  font-family: 'Playfair Display', Georgia, serif;
+  font-size: clamp(2.2rem, 4.5vw, 3.6rem);
+  font-weight: 500; line-height: 1.02; letter-spacing: -.02em;
+  margin: .5rem 0 0; color: var(--text-soft, #1a4d6e);
+  max-width: 14ch;
+}
+.ab-sus__head h2 em { font-style: italic; color: #b8541a; }
+.ab-sus__head p { margin: 0; font-size: 1.05rem; line-height: 1.7; color: var(--text, #2a3942); max-width: 54ch; }
+
+.ab-sus__grid {
+  display: grid;
+  grid-template-columns: 1.4fr 1fr;
+  gap: clamp(1.5rem, 3vw, 2.5rem);
+}
+@media (max-width: 980px) { .ab-sus__grid { grid-template-columns: 1fr; } }
+.ab-sus__pillars { display: grid; grid-template-columns: 1fr 1fr; gap: 1rem; }
+@media (max-width: 520px) { .ab-sus__pillars { grid-template-columns: 1fr; } }
+.ab-pillar {
+  background: var(--surface, #fff);
+  border: 1px solid var(--border, rgba(8,30,46,.08));
+  border-radius: 16px;
+  padding: 1.5rem 1.4rem 1.6rem;
+  display: flex; flex-direction: column; gap: .6rem;
+}
+.ab-pillar__icon {
+  width: 44px; height: 44px;
+  border-radius: 12px;
+  background: rgba(184,84,26,.12);
+  color: #b8541a;
+  display: inline-flex; align-items: center; justify-content: center;
+  margin-bottom: .35rem;
+}
+.ab-pillar h4 {
+  font-family: 'Playfair Display', Georgia, serif;
+  font-size: 1.2rem; font-weight: 500;
+  margin: 0;
+  color: var(--text-soft, #1a4d6e);
+}
+.ab-pillar p {
+  margin: 0;
+  font-size: .92rem; line-height: 1.6;
+  color: var(--text, #2a3942);
+}
+.ab-pillar__stat {
+  margin-top: auto;
+  padding-top: .85rem;
+  border-top: 1px dashed rgba(184,84,26,.3);
+  font-family: 'Playfair Display', Georgia, serif;
+  font-style: italic;
+  font-size: 1.05rem;
+  color: #b8541a;
+}
+.ab-pillar__stat strong { font-style: normal; font-weight: 500; font-size: 1.4rem; margin-right: .3rem; }
+
+.ab-sus__quote {
+  background: #f0eee4;
+  border-radius: 18px;
+  padding: 2rem 1.75rem;
+  display: flex; flex-direction: column;
+  gap: 1rem;
+  position: relative;
+}
+[data-theme="dark"] .ab-sus__quote { background: #163449; }
+.ab-sus__quote::before {
+  content: '"';
+  position: absolute; top: -.5rem; left: 1.5rem;
+  font-family: 'Playfair Display', Georgia, serif;
+  font-style: italic;
+  font-size: 8rem;
+  line-height: 1;
+  color: rgba(184,84,26,.18);
+}
+.ab-sus__quote-text {
+  font-family: 'Playfair Display', Georgia, serif;
+  font-style: italic;
+  font-size: 1.2rem;
+  line-height: 1.5;
+  color: var(--text-soft, #1a4d6e);
+  margin: 0;
+  position: relative;
+}
+.ab-sus__quote-text:first-of-type { margin-top: 1rem; }
+.ab-sus__quote-who {
+  display: flex; align-items: center; gap: .85rem;
+  padding-top: 1rem;
+  border-top: 1px solid rgba(8,30,46,.1);
+  margin-top: auto;
+}
+[data-theme="dark"] .ab-sus__quote-who { border-top-color: rgba(255,255,255,.1); }
+.ab-sus__quote-avatar {
+  width: 48px; height: 48px; border-radius: 50%;
+  background: linear-gradient(135deg, #c89878, #8a5a3a);
+  object-fit: cover;
+  object-position: center top;
+  display: inline-block;
+  flex: 0 0 auto;
+}
+.ab-sus__quote-who strong {
+  display: block;
+  font-family: var(--dp-font-sans, 'Montserrat', sans-serif);
+  font-size: .92rem; font-weight: 600;
+  color: var(--text-soft, #1a4d6e);
+}
+[data-theme="dark"] .ab-sus__quote-who strong { color: #F5F1E8; }
+.ab-sus__quote-who span {
+  font-family: var(--dp-font-sans, 'Montserrat', sans-serif);
+  font-size: 10px; letter-spacing: .14em; text-transform: uppercase;
+  color: var(--text-muted, #5a6b78);
+  font-weight: 600;
+  display: block; margin-top: 3px;
+}
+
+.ab-sus__badges {
+  margin-top: clamp(2.5rem, 5vw, 4rem);
+  padding-top: 2rem;
+  border-top: 1px solid var(--border, rgba(8,30,46,.08));
+  display: flex; gap: clamp(1.5rem, 4vw, 3.5rem); flex-wrap: wrap; align-items: center;
+  justify-content: space-between;
+}
+.ab-sus__badges-label {
+  font-family: var(--dp-font-sans, 'Montserrat', sans-serif);
+  font-size: 11px; letter-spacing: .2em; text-transform: uppercase;
+  color: var(--text-muted, #5a6b78); font-weight: 700;
+}
+.ab-sus__badge {
+  font-family: 'Playfair Display', Georgia, serif;
+  font-size: 1rem; font-style: italic;
+  color: var(--text-soft, #1a4d6e);
+  opacity: .65;
+  display: flex; align-items: center; gap: .6rem;
+}
+[data-theme="dark"] .ab-sus__badge { color: rgba(245,241,232,.7); }
+.ab-sus__badge::before {
+  content: ''; width: 6px; height: 6px; border-radius: 50%; background: #b8541a;
+}
+
