+++ src/styles/about/cta.css
@@ -1,0 +1,34 @@
+/* ============ CTA ============ */
+.ab-cta {
+  position: relative;
+  padding: clamp(5rem, 9vw, 8rem) clamp(1.25rem, 5vw, 6rem);
+  overflow: hidden;
+  isolation: isolate;
+  color: #F5F1E8;
+  text-align: center;
+}
+.ab-cta__bg { position: absolute; inset: 0; z-index: -2; }
+.ab-cta__bg img { width: 100%; height: 100%; object-fit: cover; }
+.ab-cta__bg::after {
+  content: ''; position: absolute; inset: 0;
+  background: linear-gradient(180deg, rgba(8,30,46,.55) 0%, rgba(8,30,46,.85) 100%);
+}
+.ab-cta__inner { max-width: 720px; margin: 0 auto; position: relative; }
+.ab-cta__eyebrow { display: inline-flex; justify-content: center; color: #FF6F61 !important; }
+.ab-cta__eyebrow::before { background: #FF6F61 !important; }
+.ab-cta h2 {
+  font-family: 'Playfair Display', Georgia, serif;
+  font-size: clamp(2.25rem, 4.5vw, 3.5rem);
+  font-weight: 500; line-height: 1.05; letter-spacing: -.015em;
+  margin: 1rem 0;
+  color: #F5F1E8;
+}
+.ab-cta h2 em { font-style: italic; color: #FF6F61; }
+.ab-cta p {
+  font-size: 1.05rem; line-height: 1.65;
+  color: rgba(245,241,232,.85);
+  margin: 0 auto 2rem;
+  max-width: 52ch;
+}
+.ab-cta__btns { display: inline-flex; gap: .85rem; flex-wrap: wrap; justify-content: center; }
+
