+++ src/styles/about/press.css
@@ -1,0 +1,105 @@
+/* ============ PRESS ============ */
+.ab-press {
+  padding: clamp(4rem, 8vw, 7rem) clamp(1.25rem, 5vw, 6rem);
+  max-width: 1240px; margin: 0 auto;
+}
+.ab-press__head {
+  display: grid; grid-template-columns: 1fr 1.2fr;
+  gap: clamp(2rem, 5vw, 4rem);
+  align-items: end;
+  margin-bottom: clamp(2.5rem, 5vw, 4rem);
+}
+@media (max-width: 880px) { .ab-press__head { grid-template-columns: 1fr; align-items: start; } }
+.ab-press__head h2 {
+  font-family: 'Playfair Display', Georgia, serif;
+  font-size: clamp(2.2rem, 4.5vw, 3.6rem);
+  font-weight: 500; line-height: 1.02; letter-spacing: -.02em;
+  margin: .5rem 0 0; color: var(--text-soft, #1a4d6e);
+  max-width: 13ch;
+}
+.ab-press__head h2 em { font-style: italic; color: #b8541a; }
+.ab-press__head p { margin: 0; font-size: 1.05rem; line-height: 1.7; color: var(--text, #2a3942); max-width: 54ch; }
+
+.ab-press__strip {
+  display: grid;
+  grid-template-columns: repeat(5, 1fr);
+  gap: 1px;
+  background: var(--border, rgba(8,30,46,.1));
+  border: 1px solid var(--border, rgba(8,30,46,.1));
+  border-radius: 14px;
+  overflow: hidden;
+  margin-bottom: clamp(2.5rem, 4vw, 3.5rem);
+}
+@media (max-width: 800px) { .ab-press__strip { grid-template-columns: repeat(3, 1fr); } }
+@media (max-width: 480px) { .ab-press__strip { grid-template-columns: repeat(2, 1fr); } }
+.ab-press__logo {
+  background: var(--surface, #fff);
+  padding: 1.5rem 1.25rem;
+  display: flex; align-items: center; justify-content: center;
+  font-family: 'Playfair Display', Georgia, serif;
+  font-size: 1.2rem;
+  color: var(--text-soft, #1a4d6e);
+  letter-spacing: -.005em;
+  font-weight: 500;
+  text-align: center;
+  line-height: 1.1;
+  min-height: 90px;
+  opacity: .55;
+  transition: opacity .25s;
+}
+.ab-press__logo { transition: opacity .35s var(--dp-ease-out); }
+.ab-press__logo:hover { opacity: 1; }
+.ab-press__logo small {
+  display: block;
+  font-family: var(--dp-font-sans, 'Montserrat', sans-serif);
+  font-size: 9px; letter-spacing: .18em; text-transform: uppercase;
+  color: var(--text-muted, #5a6b78); font-weight: 700;
+  margin-top: .4rem;
+}
+
+.ab-press__quotes {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 1.25rem;
+}
+@media (max-width: 880px) { .ab-press__quotes { grid-template-columns: 1fr; } }
+.ab-press__quote {
+  background: var(--surface, #fff);
+  border: 1px solid var(--border, rgba(8,30,46,.08));
+  border-radius: 16px;
+  padding: 1.75rem 1.5rem 1.5rem;
+  display: flex; flex-direction: column;
+  gap: 1rem;
+  position: relative;
+}
+.ab-press__quote-source {
+  font-family: var(--dp-font-sans, 'Montserrat', sans-serif);
+  font-size: 10px; letter-spacing: .2em; text-transform: uppercase;
+  color: #b8541a; font-weight: 700;
+}
+.ab-press__quote-text {
+  font-family: 'Playfair Display', Georgia, serif;
+  font-size: 1.15rem; line-height: 1.5;
+  color: var(--text-soft, #1a4d6e);
+  margin: 0;
+  text-wrap: pretty;
+}
+.ab-press__quote-text em { font-style: italic; color: #b8541a; }
+.ab-press__quote-foot {
+  display: flex; align-items: center; justify-content: space-between;
+  margin-top: auto;
+  padding-top: 1rem;
+  border-top: 1px solid var(--border, rgba(8,30,46,.08));
+  font-family: var(--dp-font-sans, 'Montserrat', sans-serif);
+  font-size: 11px;
+  color: var(--text-muted, #5a6b78);
+  font-weight: 600;
+}
+.ab-press__quote-foot a {
+  color: #b8541a;
+  text-decoration: none;
+  display: inline-flex; align-items: center; gap: .35rem;
+  font-weight: 600;
+}
+.ab-press__quote-foot a:hover { gap: .55rem; }
+
