+++ src/styles/about/states.css
@@ -1,0 +1,34 @@
+/* ============ DARK PATCHES ============ */
+[data-theme="dark"] .about-page { color: rgba(245,241,232,.82); }
+[data-theme="dark"] .ab-story__title,
+[data-theme="dark"] .ab-story__lead p,
+[data-theme="dark"] .ab-tl-item h4,
+[data-theme="dark"] .ab-guides__head h2,
+[data-theme="dark"] .ab-sus__head h2,
+[data-theme="dark"] .ab-press__head h2,
+[data-theme="dark"] .ab-press__quote-text,
+[data-theme="dark"] .ab-pillar h4,
+[data-theme="dark"] .ab-guide__name,
+[data-theme="dark"] .ab-sus__quote-text { color: #F5F1E8; }
+[data-theme="dark"] .ab-tl-item p,
+[data-theme="dark"] .ab-guide__bio,
+[data-theme="dark"] .ab-pillar p { color: rgba(245,241,232,.82); }
+[data-theme="dark"] .ab-guide,
+[data-theme="dark"] .ab-pillar,
+[data-theme="dark"] .ab-press__quote,
+[data-theme="dark"] .ab-press__logo { background: #112C3F; border-color: rgba(255,255,255,.08); }
+[data-theme="dark"] .ab-press__logo { color: rgba(245,241,232,.7); }
+[data-theme="dark"] .ab-press__strip { background: rgba(255,255,255,.08); border-color: rgba(255,255,255,.08); }
+[data-theme="dark"] .ab-tl-item::before { box-shadow: 0 0 0 4px #0B2030; }
+
+/* Reveal — matches homepage sections-base */
+.about-page .reveal { opacity: 0; transform: translateY(24px); transition: opacity .8s var(--dp-ease-out), transform .8s var(--dp-ease-out); }
+.about-page .reveal.is-visible { opacity: 1; transform: none; }
+
+/* Anchor sticky entrance */
+.ab-anchor { transition: background .4s ease, border-color .4s ease; }
+.ab-anchor a {
+  position: relative;
+}
+.ab-anchor a.is-active { box-shadow: 0 6px 18px rgba(184,84,26,.25); }
+
