+++ src/styles/about/story.css
@@ -1,0 +1,144 @@
+/* ============ STORY ============ */
+.ab-story {
+  padding: clamp(4rem, 8vw, 7rem) clamp(1.25rem, 5vw, 6rem);
+  max-width: 1240px;
+  margin: 0 auto;
+}
+.ab-story__head {
+  display: grid;
+  grid-template-columns: 1fr 1.4fr;
+  gap: clamp(2rem, 5vw, 5rem);
+  align-items: start;
+  margin-bottom: clamp(3rem, 6vw, 5rem);
+}
+@media (max-width: 880px) { .ab-story__head { grid-template-columns: 1fr; } }
+.ab-story__eyebrow {
+  display: inline-flex; align-items: center; gap: .55rem;
+  font-family: var(--dp-font-sans, 'Montserrat', sans-serif);
+  font-size: 11px; font-weight: 700; letter-spacing: .24em; text-transform: uppercase;
+  color: #b8541a;
+  margin-bottom: 1rem;
+}
+.ab-story__eyebrow::before { content: ''; width: 18px; height: 1px; background: #b8541a; }
+.ab-story__title {
+  font-family: 'Playfair Display', Georgia, serif;
+  font-size: clamp(2.4rem, 5vw, 4rem);
+  font-weight: 500;
+  line-height: 1.02;
+  letter-spacing: -.02em;
+  margin: 0;
+  color: var(--text-soft, #1a4d6e);
+  max-width: 13ch;
+}
+.ab-story__title em { font-style: italic; color: #b8541a; }
+.ab-story__lead p {
+  font-family: 'Playfair Display', Georgia, serif;
+  font-style: italic;
+  font-size: clamp(1.15rem, 1.6vw, 1.45rem);
+  line-height: 1.55;
+  color: var(--text-soft, #1a4d6e);
+  margin: 0 0 1.25rem;
+  text-wrap: pretty;
+}
+.ab-story__lead p + p {
+  font-style: normal;
+  font-family: var(--dp-font-sans, 'Montserrat', sans-serif);
+  font-size: 1.02rem;
+  color: var(--text, #2a3942);
+  line-height: 1.75;
+}
+
+/* Timeline */
+.ab-timeline {
+  display: grid;
+  grid-template-columns: 1.1fr 1fr;
+  gap: clamp(2rem, 5vw, 5rem);
+  align-items: start;
+}
+@media (max-width: 980px) { .ab-timeline { grid-template-columns: 1fr; } }
+.ab-timeline__rail {
+  position: relative;
+  padding-left: 2rem;
+}
+.ab-timeline__rail::before {
+  content: '';
+  position: absolute; left: 7px; top: 12px; bottom: 12px;
+  width: 2px;
+  background: linear-gradient(180deg, #b8541a 0%, rgba(184,84,26,.15) 100%);
+  border-radius: 1px;
+}
+.ab-tl-item {
+  position: relative;
+  padding-bottom: 2.25rem;
+}
+.ab-tl-item:last-child { padding-bottom: 0; }
+.ab-tl-item::before {
+  content: '';
+  position: absolute; left: -2rem; top: 7px;
+  width: 16px; height: 16px;
+  border-radius: 50%;
+  background: var(--surface, #fff);
+  border: 2px solid #b8541a;
+  box-shadow: 0 0 0 4px var(--page-bg, #FFFCF5);
+}
+.ab-tl-item__year {
+  font-family: 'Playfair Display', Georgia, serif;
+  font-style: italic;
+  font-size: 1.05rem;
+  color: #b8541a;
+  letter-spacing: .04em;
+  margin-bottom: .35rem;
+}
+.ab-tl-item h4 {
+  font-family: 'Playfair Display', Georgia, serif;
+  font-size: 1.35rem;
+  font-weight: 500;
+  margin: 0 0 .5rem;
+  color: var(--text-soft, #1a4d6e);
+  line-height: 1.2;
+}
+.ab-tl-item p {
+  margin: 0;
+  font-size: .98rem;
+  line-height: 1.65;
+  color: var(--text, #2a3942);
+  max-width: 50ch;
+}
+.ab-timeline__photo {
+  position: relative;
+  border-radius: 18px;
+  overflow: hidden;
+  aspect-ratio: 4/5;
+  background: #2a1a10;
+}
+@media (min-width: 981px) {
+  .ab-timeline__photo-wrap { align-self: stretch; position: relative; }
+  .ab-timeline__photo-wrap .ab-timeline__photo {
+    position: sticky;
+    top: calc(var(--nav-height, 66px) + 1.5rem);
+    aspect-ratio: auto;
+    height: min(64vh, 560px);
+  }
+  .ab-timeline__photo-wrap.js-pin .ab-timeline__photo {
+    position: relative;
+    top: auto;
+    transform: translate3d(0, var(--pin-y, 0px), 0);
+    will-change: transform;
+    transition: none;
+  }
+}
+.ab-timeline__photo img { width: 100%; height: 100%; object-fit: cover; display: block; }
+.ab-timeline__caption {
+  position: absolute; left: 1.25rem; bottom: 1.25rem; right: 1.25rem; z-index: 2;
+  color: #F5F1E8;
+  font-family: 'Playfair Display', Georgia, serif;
+  font-style: italic;
+  font-size: 1rem;
+  line-height: 1.45;
+  text-shadow: 0 2px 14px rgba(0,0,0,.5);
+}
+.ab-timeline__photo::after {
+  content: ''; position: absolute; inset: 0;
+  background: linear-gradient(180deg, transparent 50%, rgba(20,12,8,.7) 100%);
+}
+
